Modify engine to serialize data as EDN
Previously we were relying on Postgres JSONB to serialize Backtick job
data. This is not ideal because JSON is lower-fidelity than Clojure. For
example, passing an argument map with keyword keys, when deserialized
would have string keys. EDN on the other hand, is pretty much the same
thing as Clojure, so no loss of fidelity occurs.