Backtick migrations conflict with other Clams projects'
Clams migration system doesn't expect a library to drop resources under
the migrations/ prefix. And the JVM is not smart enough to contain
packages' resources under distinct namespaces. To avoid the conflict,
this diff moves Backtick's migrations under a separate prefix. Although
this solves the immediate issue, it's still annoying because the library
user will have to manually copy the migrations scripts under their own
migrations prefix (or system, if not Clams'). Also modifies the Backtick
makefile to symlink the migrations directory so unit tests will still
run as expected.