Merge priority and run_at
It used to be that priority was used when a job was created and when
it was retried.  And then purely as the ordering of jobs.  Now we wish
to only retry on a specific time schedule, so the only time when we
would want to run a job ASAP would be when the initial job is created.

We can accomplish this by getting rid of run_at and only running jobs
when priority >= now.

I have a few more changes I'd like to make, but wanted to get some
feedback before we continued.