Allow jobs to be canceled before they are executed
Adds a new `canceled` state. Jobs that are set as `canceled` will not be
dequeued for execution and will eventually get cleaned up by the existing
cleaner. This shouldn't generally be necessary but the ability to drain the
queue and just skip all its jobs can be useful for debugging or error recovery
in a running queue (where it's not practical to go manually poking around the database).