Fix function unquote in core macros
A function name was not being passed correctly in the core macros
`define-recurring` and `define-cron`. Previously macroexpansion generated code
like this:

```
user=> (macroexpand '(bt/define-recurring testf 999 [] (println ">>> this is it")))
(do
 (clojure.core/defn testf [] (println ">>> this is it"))
  (backtick.registry/register "user/testf" testf)
  (#<Fn@70a5db0d backtick.core/schedule_recurring> 999 testf))
```

Notice that crazy `#<Fn@xxxxxxxx` object? Inexplicably this seems to have been
functioning correctly, but in the interests of readable code, and safety, this
commit fixes it so the correct function _symbol_ is generated there instead.

Additionally unit test coverage has been added for these macros, and a more
informative error message has been added for when the database URL is unset.